Split up E2E test files
